Add SlotStakeCounters.
